progress counter is a separate header
